Wrap createOrder in transaction; move helper functions into helpers folder
