+++ server/src/templates/member/orderConfirmationEmail.html
@@ -142,7 +142,7 @@
       <table class="details">
         {{#each pizzas}}
         <tr>
-          <th colspan="3">Pizza #{{number}}</th>
+          <th colspan="3">Pizza #{{index}}</th>
         </tr>
         <tr>
           <td>Size</td>
